test(processors): move tests to 'medium' and 'low' folders
