+++ src/shared/processors/low/__tests__/gallery.tests.js
@@ -1,6 +1,6 @@
 /* eslint-disable react/prop-types */
-import Gallery from '../../components/Gallery';
-import processor from '../low/gallery';
+import Gallery from '../../../components/Gallery';
+import processor from '../gallery';
 
 const gallery = {
   type: 'element',
